backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
requests==2.31.0
pydantic==2.5.0